+++ docs/ChangeLog.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve sanity checks in the **Message** class.
 - **Message::clear** method preserves reference-count and ttl.
-- Deprecate **Message::getHeader** and **Message::getStorage** methods.
+- Rename **Message::getHeader** method to **Message::header**.
+- Rename **Message::getStorage** method to **Message::buffer**.
 
 ## [5.0.0] - 2021-11-18